database-setup/pyproject.toml
[project]
name = "database-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
